requirements.txt
colorama==0.4.6
greenlet==3.2.2
pretty-errors==1.2.25
py4j==0.10.9.9
pyspark==4.0.0
python-dotenv==1.1.0
SQLAlchemy==2.0.41
typing_extensions==4.13.2
